+++ eslint.config.mjs
@@ -15,7 +15,7 @@ export default [
   ...nx.configs['flat/typescript'],
   ...nx.configs['flat/javascript'],
   {
-    ignores: ['**/dist', '**/vitest.config.*.timestamp*']
+    ignores: ['**/dist', '**/vitest.config.*.timestamp*', '**/*.generated.*']
   },
   {
     plugins: {